+++ packages/core-js-compat/src/modules-by-versions.mjs
@@ -323,4 +323,9 @@ export default {
     'es.weak-map.get-or-insert',
     'es.weak-map.get-or-insert-computed',
   ],
+  3.49: [
+    'web.request-idle-callback',
+    'web.cancel-idle-callback',
+    'web.idle-deadline.constructor',
+  ],
 };