+++ packages/core-js/web/index.js
@@ -19,6 +19,9 @@ require('../modules/web.url-search-params');
 require('../modules/web.url-search-params.delete');
 require('../modules/web.url-search-params.has');
 require('../modules/web.url-search-params.size');
+require('../modules/web.request-idle-callback');
+require('../modules/web.cancel-idle-callback');
+require('../modules/web.idle-deadline.constructor');
 var path = require('../internals/path');
 
 module.exports = path;