+++ tests/compat/tests.js
@@ -326,6 +326,10 @@ function TIMERS() {
   })());
 }
 
+function IDLE_CALLBACKS() {
+  return requestIdleCallback && cancelIdleCallback && IdleDeadline;
+}
+
 // https://github.com/tc39/ecma262/pull/3467
 function checkIteratorClosingOnEarlyError(METHOD_NAME, ExpectedError) {
   return function () {
@@ -2275,5 +2279,16 @@ GLOBAL.tests = {
   }],
   'web.url-search-params.size': [URL_AND_URL_SEARCH_PARAMS_SUPPORT, function () {
     return 'size' in URLSearchParams.prototype;
-  }]
+  }],
+  'web.request-idle-callback': [
+    IDLE_CALLBACKS,
+    function () {
+      // Firefox bug on negative / 0 timeout
+      // Proper test would be requestIdleCallback(a=>(console.log(a.didTimeout)), {timeout: -1})
+      // -- if that gives false we're good
+      return /Firefox .\./.test(USERAGENT);
+    }
+  ],
+  'web.cancel-idle-callback': IDLE_CALLBACKS,
+  'web.idle-deadline.constructor': IDLE_CALLBACKS
 };