+++ tests/eslint/fix-runner.mjs
@@ -1,0 +1,3 @@
+process.env.ESLINT_USE_FLAT_CONFIG = true;
+
+await $`TIMING=1 eslint --concurrency=auto --fix --config ./tests/eslint/eslint.config.js ./`;